+++ packages/components-react/src/routes.ts
@@ -313,6 +313,11 @@ export const routes: RouteType[] = [
     component: fromExamples.TableExampleAdvancedPage,
   },
   {
+    name: 'Tabs Bar Example',
+    path: '/tabs-bar-example',
+    component: fromExamples.TabsBarExamplePage,
+  },
+  {
     name: 'Toast Example',
     path: '/toast-example',
     component: fromExamples.ToastExamplePage,