+++ packages/components/CHANGELOG.md
@@ -8,7 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
-#### Added 
+#### Added
+- `:focus-visible` Content of selected Tab in `Tabs` component gets focus styling
+- improved accessibility of `Text Field Wrapper` and `Textarea Wrapper` when `maxlength` attribute is set
 - `Modal` aria property
 - `Modal` class for slotted elements to make content full-width